+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/controller/ChallengeController.java
@@ -1,6 +1,6 @@
 package com.itachallenge.challenge.controller;
 
-import com.itachallenge.challenge.dtos.ChallengeDto;
+import com.itachallenge.challenge.dto.ChallengeDto;
 import com.itachallenge.challenge.service.ChallengeService;
 import io.swagger.v3.oas.annotations.Operation;
 import org.slf4j.Logger;
