+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/helper/ModelMapperConverters.java
@@ -1,7 +1,7 @@
 package com.itachallenge.challenge.helper;
 
-import com.itachallenge.challenge.documents.Challenge;
-import com.itachallenge.challenge.dtos.ChallengeDto;
+import com.itachallenge.challenge.document.Challenge;
+import com.itachallenge.challenge.dto.ChallengeDto;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
